basic filtering
We'll need some filtering functionality at one point, so let's prepare
for it. There are three things to note:

1. We are so far parsing all the inputs within each chunk's filtering
call (instead of parsing it at once and passing it in), so that we can
avoid interface{}. Perf might get to us later.
2. We had to rework the Bitmap library (for the better) and add support
for inverting the whole bitmap. I'm glad it's there now.
3. There is only a handful of tests, we need to add more.