basic support for query limits
We need to handle query limits in two parts of the query - if we have a
filter in place, we need to prune that one. If we don't have a filter,
we need to prune in a different way.

The implementation is pretty bad and untested, we should improve upon
it.