pruning bools fixed
Pruning of bools didn't work right, because we didn't track what row we
were on. And because we didn't have correct tests, we did not quite
catch it in time.

We had to fix all the pruning tests, which uncovered one more subtle bug
(nullability bitmap wasn't set right during pruning).