force binding ports to ensure repeatable builds
If the requested port is busy for any reason, smda will try to bind a
different port. However, this is not desirable in some scenarios, where
we'd rather fail than bind a new port (e.g. we're proxying data to a
given port, so we need the app on this very port).

Unfortunately, this is not very well testable with the current setup,
we'll need to figure out a different architecture for this. There is a
sketch of a test, but it's commented out for now.