integrity checks when reading data
We write CRC32 checksums for all our columns into our binary
representations and then check these upon reading. It creates a bit of
overhead (in terms of allocations, the perf loss is negligible).

No tests atm, but they should be fairly straightforward - flip a bit and
watch the deserialiser burn.