moving files to sub packages
This should clean things up, we'll also be more explicit about exported
types and functions.

Some functionality had to be altered slightly, as we cannot add
receivers to types outside our packages.

Also, database should be further broken down, but it was hard to do so,
as there were a lot of circular imports.