moving typed columns/chunks into a subpackage
Nearly all files touched, only one breaking thing (we cannot access
chunk's unexported fields, so we cannot e.g. check for nullability).

As circular imports are not allowed, we had to split out
inference_types.go and move a part of it into column/schema and leave a
part in it.

These package splits are fairly painful, but will lead to smaller
package surfaces, which should aid development, testing etc.