basic literals implementation
Literals are column chunks, which only contain a single value. These can
be useful in a number of ways:
- we can convert string/int/float/bool literals in expressions into
chunks without taking up much memory, but still conforming to the same
interface as "fat" chunks. This simplifies their usage.
- it's an extreme form of RLE - this could be useful in the future, if
we decide to parse columns into literals first and only convert to "fat"
chunks once we detect a second distinct value (you could then have a
column, that will be a literal in some chunks and not others)

At this point, the implementation is only partial - we don't have any
binary marshaling, because we only use them in expression evaluation.