chunk creation doesn't allocate a nullability bitmap
We'd always create a new bitmap for nullability - even when we knew we
wouldn't have any NULLs. The new default is not to create it, but we
have to be rather cautious, because any use of that nullability bitmap
will result in a segfault.

We've already introduced a refactored bitmap.Serialize to take care of
nils, but we'll have to introduce this nilness check more widely.