dist: build zip files instead of tar.gz
While tar.gz are fairly standard in terms of publishing artifacts, this
tool is for broader use than just build/release automation, so end users
might not have tar installed and/or might not know how to use it. And
making it difficult to take the very first step in using this piece of
software would be quite suboptimal.