Basic support for the Lambda function runtime
We had to rehash quite a few things to make this work, but now you can
run the runtime within Lambda. At this point it's super stateless - once
the Lambda container gets (re)created, you lose all the data, which is
obviously not great, but we won't have state support until we figure out
our tiered storage story.

In the meantime, there's a poor man's Terraform script to deploy the
runtime to Lambda, utilising its new Function URLs support. So all you
need is a bit of IAM and Lambda, no API Gateways or other products.